WIP: Add Initial Work for Custom Select Element
Originally, when we started writing this component,
we wanted to give an example of making the component
_accessible_ and using semantic HTML. However, we've
kind of been brought to a halt on this because
apparenty a11y gets VERY complicated when it comes
to ARIA attributes that try to reach for elements
across the ShadowBoundary. We don't really want to
present people with something that's inaccessible,
because it won't be good for their users and it won't
give them good solutions or best practices. So...
for now we're stashing our work. If we can come up
with something that makes the Custom Select Component
accessible, we'll return to our efforts here. Until
then, we'll probably do some cleanup in the SuperTokens
examples or in the package that we have not yet released.

For anyone looking at this code, please bear in mind
that it is incomplete. Feature-wise, it's probably
at least 75% done if you wanted to take it for a spin.
If you don't care about a11y, it is still very possible
and easy to finish building out this component.

Future Note: Can't use `onblur` for the combobox. We need
to use a different kind of listener so that users can
select options by clicking them.

For A11y Issues, See:
- https://nolanlawson.com/2022/11/28/shadow-dom-and-accessibility-the-trouble-with-aria/
- https://github.com/whatwg/html/issues/3219