WIP: Add Additional Work to `CustomSelect` before Trying Light DOM
This is the work that we currently have for `CustomSelect`.
We are going to try putting everything in the light DOM
instead of the Shadow DOM in order to get a11y working...
Shadow DOM was great while it lasted, but it might not
be the way that we need to go...

We're saving this commit so that it's clear where our
progress was before we pivoted. And we're doing this,
again, because of a11y issues with the Shadow DOM.

This commit has no commits about the improvements that were
made. You'll have to sift through those yourself (sorry).

Advances the Following WIPs:
- afa395986bad91f0fed3c445cfa8e15615c1dae0