refactor: Use Vite for Rapid Testing
There comes a point where `npx tsc --watch` just doesn't
cut it. Using `vite` will make life easier for us. And
with this approach, we hopefully won't have to worry
as much about the order in which certain things happen
(e.g., the order in which certain scripts are loaded
into the DOM, now that we only have 1 script tag).