chore: Start Adding Playwright Tests for Combobox Component
We introduced the mouse tests, the `ArrowDown` tests,
and the `End` tests. There are more to come... We are
basing the tests off the WAI ARIA Spec instead of
trying to come up with our own ideas -- which makes
testing much easier and more more reliable.

NOTE: When we installed Playwright, we left their
setup along for the most part. We added their
`.gitignore` updates. We accepted their
`playwright.config.ts` file (with an edit
to the kinds of files/folders that should
be searched for tests). We accepted
their `.yml` for testing in GitHub CI (which
probably won't work for right now). And that's
about it. ... Just something to keep in mind if
you ever want to revert to the defaults. :)

Another Note: We made slight updates to the
`ComboboxContainer` to help with testing.
We also changed `Combobox.css` to an SCSS file
to help with cross-browser testing.