fix: Transfer `ComboboxContainer` Attributes to `ComboboxField`
This is certainly an unorthodox way of doing things, but
the `CombobxContainer` serves no other purposes except
to contain the elements necessary for making the
`combobox` widget work. We can change the approach in
the future if needed... But in order to get the user
experience that we desire, we'll probably have to do
something of a similar degree of hacking.