chore: Switch to `MIT` License
The goal of the GPL v3 license was to make sure
that all efforts related to Web Components that
were inspired by this repo would remain public
so that other developers could continue to benefit
from any additional iterations.

Unfortunately, it seems like the GPL v3 might be a
little stricter than what we wanted. What we wanted
was something that would allow companies with
private applications to use our components privately.
But if they were to offer the components (perhaps
as some UI library that's paid for), then they would
have to make the code open source.

My understanding is that -- under the GPL v3 -- when
a work is conveyed to a user, the work must be made
open source. Unfortunately, Web Applications _always_
convey client side JavaScript to their users. The
goal isn't to require people to share the codebase of
their entire application. And the goal isn't to
require any modifications to the Web Component examples
to be public either. The goal is only to protect end
developers from people who would want to try
to sell another version of these Web Components,
specifically. We haven't found a license for this yet.
If one doesn't exist... well, we'll stick to the MIT.
It's (hopefully) unlikely that anyone would try to
re-sell the work that I've done here while keeping
said work hidden from end developers.