refactor: Update `ComboboxOption` to Act More Like Element
The new functionality more closely matches up with
what would be expected of the `<select>` element.
The one caveat to this regarding the `defaultSelected`
property (or more specifically, the `selected` attribute)
is that its functionality is not disabled once the
`combobox` value is modified. (We're foregoing this
functionality since we don't currently know how to
implement it without leaking implementation details,
and it's a rather obscure "use case".)

We added tests for the properties/attributes
related to the `selected` property since the
logic surrounding it is so complex.