refactor: Use Raw CSS for Styles
Now that CSS Nesting is supported in modern
browsers (albeit with the need for the `&`,
except in the case of FireFox), it seems
best to just use regular CSS. The VS Code
IntelliSense is a little worse with this
approach, but we get to skip a build step
now. So overall it's a big win. Hopefully
VS Code catches up.