refactor: Rename `ComboboxContainer` to `SelectEnhancer`
In order to better support SSR and/or users without access to JavaScript,
we're changing our approach to our Combobox Web Component. Instead of
having devs render unrecognized elements to the DOM, we have them
render the recognized `<select>` element to the DOM first. Then we use
that element to initialize our fully-packaged Combobox component. This
guarantees that users without access to JS will still be presented with
a form control that works correctly. Additionally, it might make
initialization less complicated (since no customized DOM elements will
be connected to the DOM and/or initialized until we're done setting
everything up).

This new logic is NOT present in this commit. Instead, this commit
renames the wrapping Custom Element to `SelectEnhancer` to better
represent what the Web Component will be doing. We'll add the related
logic later (perhas in the next commit. Maybe we might even squash said
commit into this one).