refactor: Display `listbox` based on `combobox` expansion, not JavaScript
This ties the Visual User experience in more closely with a valid
Screen Reader experience. This is what we want! If the accessibility
of our component is broken, then the visuals of our component should
also be broken (making catching/fixing a11y bugs significantly easier).
This approach has no disadvantages, so it makes the most sense to take.