refactor: Make Combobox component something that enhances <select>s
The problem with having a Combobox component that requires JavaScript
is that it will never work for users lacking JavaScript. By creating
a Web Component that _wraps_ (and replaces) as <select> element, we
guarantee that users will have a "regular combobox experience" when
JS is unavailable, and that they'll have a _superior_ combobox
experience when JS _is_ available. All that's needed to get this
working as desired is to improve the CSS for all <select>s that appear
within a <select-enhancer>. Then, users won't ever be able to tell
that the <select> was replaced underneath the hood.

(CSS still needs to be added.)