chore: Upgrade GitHub Actions
We might run into issues with artifact names from now on.
If that's the case, I don't think I'll fret about it too
much as long as the tests are passing. We can always delete
artifacts having the same name, or we can randomly generate
artifact names. Not the biggest deal.