refactor: Enable `ComboboxField.value` to be `string | null`
This allows us (and other developers) to identify when the `combobox`
has not yet been initialized (e.g., if it wasn't given any `option`s).
This change will ultimately make our code much simpler to work with.

Note that this change not only includes `ComboboxField.value`, but the
value that it exposes to its owning `<form>` as well. (Of course, it
should be expected that a form control's `value` matches the value it
will send to the server through its owning form.)