chore: Add (simple) Performance Testing for event handler management
Our components often use "cached" `static` event handlers by
default. And when we _need_ to have access to `this`, we leverage
arrow functions that are created and assigned at construction.
We wanted to use performance testing to see if our reasoning
for doing this was justified, and it indeed seems justified.

As one would expect, memory usage improves dramatically if you only
create a function once for event handling and you simply reference
that function N times (instead of creating it N times).

Memory is worse if you create the function (e.g., an arrow function)
every time the component is constructed. (Simulated by assigning
values to an array.) Interestingly enough -- though not surprisingly --
memory consumption with `classInstance.cachedMethod.bind(classInstance)`
is better than memory consumption with arrow functions created on
every class construction -- even though BOTH approaches result in new
functions getting created every time a class is instantiated. This
makes sense, considering the fact that one would expect a reference to
the original function to be "maintained" and "reused" to some extent,
and for `this` to be "passed in" to said function -- at least in some sense.
However, the memory gain here from our test was only ~15%, and that was
for a _very specific_ test. So, for the sake of maintainability, this method
is probably not worth employing in our codebase unless it would truly resolve
bottlenecking issues for developers (highly unlikely to be what devs need
if they start facing bottlenecking issues).