fix: Improve `option` Search performance and always exclude "EmptyOption"
It seems like it's faster to loop through an element's children by
traversing the "Linked List Relationship" rather than the
"Array Relationship". This certainly seems to be the case from our
benchmarks.

Additionally, in the past, it seems like there were cases where we
accidentally included the "EmptyOption" in the Matching Options results.
That bug should be fixed now, but we should add tests to verify that
the fix genuinely works.