refactor: Skip preliminary check for newlines in `#handleSearch`
Almost certainly, `String.prototype.includes` has a time complexity
of O(N), where N is the number of characters in the string. That is,
the function would have to visit every character in the string until
it finds a newline. If it finds no newline, the function returns
`false`.

Practically speaking, most of the time people will be providing
normal characters, not newlines. This means that we'd be doing
useless, worst-case-scenario O(N) checks most of the time.

Additionally, although we don't know how `replaceAll` and RegExp
explicitly work under the hood, one could imagine that in worst
case scenarios, they will have to visit every character in the
string since we're talking about global character replacements.

If we're going to have to do an O(N) operation anyway, why bother
with a preliminary one? Although we haven't done any measurements,
doesn't it seem like it would be more performant to just do the
replacement call right off the bat instead of doubling up on
O(N) (or higher) time losses?

This was effectively our reasoning for the refactor. We can do
time measurements later. But just thinking about it, our
conclusions here are likely accurate/right.