chore: Minor updates to notes/comments. Make `#emptyOption` a `span`.
The only functional update is changing the `#emptyOption` from a `div`
to a `span`. This is _just in case_ someone wants to customize the
wrapping `[role="listbox"]`, and they use an Inline Element for it
instead of a Block Element. The `<span>` element will fit most
use cases -- certainly more than the `<div>` element. And the desired
layout can still be achieved by changing the `display` of the `<span>`.

Everything else is just an update to notes/comments/todos (or
_spacing_ between lines of code).