chore: Add Tests for new Filter Mode Features (User Interactions)
To the best of our current ability, this work rigorously tests the
various filter modes (`unclearable`, `clearable`, and `anyvalue`).
This commit PRIMARILY addresses the tests involving User Interactions.
In fact, we would consider the User Interaction tests "complete" now.

Tests that are more API-leaning (e.g., testing what happens when
a developer changes a property on one of the Custom Elements) will
be added next. (There's so much to test that it seemed better to
split the tests into these 2 groups).

Note that right now, all tests are passing. This means the features
for non-filter mode are behaving as expected. It also means the
features for filter mode are proven to be "backwards compatible",
so to speak, if the `combobox` is configured with
`[filteris="unclearable"]`. Of course, we want to prove that everything
works rightly regardldess of the filter mode. This will be accomplished
by completing all of the API Tests.

Additional Component Changes:
- In `filter` mode, default filtering has switched to `startsWith()`.
  This results in a more consistent UX whether the `filter` is turned
  on or off. Additionally, it probably helps users narrow down to their
  desired option(s) more easily. (They know everything that starts with
  "A", then "Ap", instead of being thrown a random slop of options that
  have inclusive letters.) We might want to update our Development Notes
  regarding this decision.
- The "No Matches" element is explicitly marked with `aria-hidden` now
  for the sake of Playwright tests (and potentially Screen Readers).
- Fixed the `disabled` state of the `combobox` to behave correctly in
  `filter` mode across all browsers. (Previously, it only worked in Chrome.)
- Using CSS, the `combobox` no longer shows the "No Matches" element
  (or the `listbox`) if `filteris="anyvalue"`.

Additional Test Changes:
- Added a `renderHTMLToPage` function for rendering raw HTML to `page`s
  in Playwright. This new approach is MUCH more convenient for dynamic
  HTML.
- Added a custom assertion for verifying where the cursor/`Selection`
  is for a `combobox`.
- Made `associateComboboxWithForm` a global helper function.
- The `expectOptionToBe(Active|Selected)` methods now use the `exact`
  option when using `page.getByRole("option")`. We found this to be
  necessary in cases where multiple `option`s started with the same
  string.
- The new `getLocationOf()` helper can now be used as a way to easily
  click between characters in a `Text` Node.
- Updated some tests to use `Locator.press()`/`Page.keyboard.press()`
  instead of `Locator.fill()`. This is because of a bug that appears in
  Playwright when it's run against Firefox (at least on MacOS). More
  details of this Playwright bug can be found in `NOTES.md` or
  microsoft/playwright#36715.