refactor: Use `renderPageToHTML` more and Add `getFilterAttrs` helper
The `getFilterAttrs` helper will help us avoid some boiler plate when
it comes to writing tests that can switch between "Regular" mode and
"Filterable" mode.

With how complex our tests are getting, `renderPageToHTML` is becoming
borderline mandatory for any tests that need to render their own HTML.
It's just a huge inconvenience to have to pass tons of variables to
a `page.evaluate` function directly. Much better to just pass the
generated HTML template. We should aim to use this function in all
tests as we continue our test updates to support filter mode.

NOTE: We discovered that many of the API tests that render their own
HTML actually AREN'T testing anything in Filter Mode (because the
rendered HTML isn't using anything like `getFilterAttrs`). So we
actually don't have the guarantee of "backwards compatibility" yet...
But hopefully we will when we update all of our tests.