chore: Minor CI / Test Updates
After starting down the path of guaranteeing that all of our tests are
executed in "Regular" mode AND in "Filterable" mode, we discovered that
for some reason, the `<dialog>`-closing test fails on Linux in CI when
Playwright runs against `WebKit`. This issue also happens on our
physical Linux machine. See: microsoft/playwright#36727.

No problems were encountered when all tests were run locally (in UI
mode or in headless mode) on MacOS. So we tried running the tests again
in CI on MacOS. With that setup, all tests passed (though we had to change
another `Locator.fill()` to `Page.keyboard.press()` to resolve more
Firefox issues). Since practically speaking, Safari can only be used
on MacOS, we consider all of our tests to be passing.

Unfortunately, the cost for running MacOS on GitHub Actions is incredibly
high compared to running Linux. In fact, as of 2025-07-19, MacOS costs
10x as much as Linux (and therefore eats 10x as many runner minutes).
Consequently, we'll be changing our OS back to Linux, which means that
the 1 test will be failing for a while until something on the Operating
System side of things (or on Playwright's side of things) gets fixed.

See: https://docs.github.com/en/billing/managing-billing-for-your-products/about-billing-for-github-actions#minute-multipliers
for info on GitHub Action's multiplications of minutes on various Operating Systems.

Note: We made slight updates to the `<dialog>`-closing test since
it seemed fitting.

When the project is more mature, we may consider splitting our Playwright
tests in CI between MacOS (for WebKit) and Linux (for everything else).
This sounds like a more reliable model; but for now, the incurred costs
are not worth it. I guess there will come a time when we push to `main`
far less often because the `Combobox` component will be considered
"complete" (though more components might come later). Perhaps at that
point such a model will be less risky.