chore: Add tests for supported `ComboboxField` methods
When we say, "Supported `ComboboxField` Methods", we mean those methods
which we guarantee to the public should be functional. These are distinct
from the Validation Methods, which are basically just hooks into the
`ElementInternals` methods anyway.

Methods which are effectively internals that we don't mind exposing
(like `getOptionByValue`) or which should only ever be triggered
by User Interactions or internally (like the lifecycle/form callbacks)
will not be tested. Library Consumers should avoid using these directly.