chore: Add support for `setCustomValidity()` and `valuemissingerror`
Originally, the gameplan was just to add support for `setCustomValidity()`.
This would give developers greater flexibility with error messaging, and
it would make our `<combobox-field>` integrate more seamlessly with our
`FormValidityObserver` in the `@form-observer/core` library.

However, we recently learned that `ElementInternals.setValidity()` is VERY
inconvenient to work with. It was so inconvenient that it was easier to
just implement the `valuemissingerror` attribute at the same time.
(This attribute controls the error displayed to users when the `required`
constraint of the component is broken.)

It might be good to open a WHATWG GitHub issue regarding this. But in the
interest of time, we'll just stick to working on this component for now.