chore: Correct `ComboboxField`'s event dispatching logic and add tests
This commit adds support for dispatching the `input` and change` events whenever
an update to the `combobox`'s filter causes an update to its value. There were
apparently bugs in the previous implementation, so those have been fixed.
Additionally, the `#handleSearch` method has been updated to only run when the
`beforeinput` event `isTrusted` (i.e., when the event is triggered by a
User Interaction). All of these updates cause our component to behave more
like the `<input>` element when in `anyvalue` mode (and in some cases, when
in `clearable` mode).

As an added change, TypeScript was updated. (We did this because of some
debugging we were doing with our unique `Symbol`s approach.) This required
2 minor changes in the `ComboboxField` file. Those changes are included here.