chore: Minor improvements to `ComboboxOption` code/comments
There are no real functional changes here. These are just changes
that help the maintainability of the component and/or help with
documentation for consumers.

The choice to default the `value` getter to the `label` getter
instead of the `textContent` property was done in case we decide
to make some future decisions. Currently, however, the change
is inconsequential since the `label` points to the `textContent`.