feat: Enable developers to COMPLETELY override filtering logic/perf
The `optionMatchesFilter()` method could be overridden to change the
filtering logic. However, this method alone could not be used to
tune the performance of the filtering logic. For example, some people
may prefer to filter the `option`s with a `Trie` (which is much more
performant than traversing an ArrayList). Still, other people may load
the `option`s asynchronously and not want any filtering logic to run on
the client side at all. This new, overridable method gives developers
the means to accomplish this. We don't think it's perfect, but we do
think it's "ergonomic enough", given our design constraints.

If new improvements or ideas come in the future, we'll explore them.
But this should be clear and accessible enough for the time being --
for those who are _unusually_ passionate about performance.