fix: Don't highlight Combobox Text when disconnected from the DOM
This was a bug that existed during transitions from non-filter mode
to filter mode. `DocumentFragment`s don't have an `activeElement`,
so this logic was useless in cases where the `combobox`'s attributes
were being changed outside of the `Document`.

Although we're calling this a "bug", it technically couldn't result
in any thrown errors. Nonetheless, the new implementation is more
representative of what we're after. And this isn't wasteful since
we already had a `this.isConnected` check in place. (If we didn't
have said check in place already, we may not have made this refactor.)