refactor: Simplify the way that the "No Matches Message" is handled
In the previous implementation, the "No Matches Message" was a chunk of text
placed inside a non-`[role="option"]` element. This element was appended to
the `listbox` when the User needed to see the message, but it was hidden from
the Accessibility Tree so that it wouldn't be flagged as invalid. (This seemed
appropriate to do since the message is also irrelevant to Screen Reader Users,
who will be told if a `listbox` is empty upon visiting it.)

Ultimately, we found this approach to be too complicated. When it came to
looping through the `option`s or running our dynamic `option` handling logic,
the approach of appending a non-`option` element to the `listbox` introduced
too much complexity to be worth maintaining. So we decided to take a different
approach where a literal element is NEVER appended to the `listbox`. Instead,
`::after`, which is not a "true" element, is appended as needed to show the
"No Matches Message". And this message is still hidden from Screen Readers.
There are other ways to accomplish this that we rejected. The reasoning for
that will be expressed separately -- ideally in a Markdown file.

Changes:
- DO NOT hide the "No Matches Message" from Users in `anyvalue` mode. Some
  developers might want to display this message in `anyvalue` mode. It should
  be up to them whether/when they want to hide it.
- Add `visibility: hidden` to the filtered-out `option` styles. For some reason,
  this is necessary for the "Safari + VoiceOver" combination to understand the
  TRUE number of `listbox` > `option`s that are available for navigation and/or
  selection. This is probably a bug in Apple's tooling?
- Use `listbox::after` to display the "No Matches Message" instead of a literal
  HTMLElement that is accessibly hidden from the DOM. This simplifies A LOT of
  code, though it unfortunately also introduces some slightly-unusual logic to
  our codebase.
  - Note that in order to change the content of `listbox::after`, we're having it
    point to a `[nomatchesmessage]` attribute in a unique and perhaps-unorthodox
    way. (Hopefully this isn't _too_ complicated though. And the end DX is better
    than some alternatives, we believe.)
- Make the default "No Matches Message" customizable. This will likely save a lot
  of text/boilerplate, especially for internationalized applications.
- FIX: Don't remove the "No Matches Message" when the `combobox` is collapsed if
  it was collapsed without any `option`s.
- Create a `ComboboxListbox` component. As more behavior/expectations get added
  to the `listbox` element, we're starting to conclude that it would be better
  for devs to receive a "fully-packaged" `listbox` (rather than devs having to
  create `listbox`es on their own). This `listbox` doesn't do much, but it will
  likely help reduce confusion/overhead for developers.
  - `ComboboxField.listbox` must now point to a `ComboboxListbox`, in addition to
    pointing to a `[role="listbox"]`.
- Minor code simplifications (especially in `SelectEnhancer`).
- Updated/Corrected comments. Added some new TODO items to code.

Additional Fixes/Changes:
- Cause `combobox` to scroll horizontally (like an `<input>`) if it has too
  much text.
- Added an `expect(combobox).toShowNoMatchesMessage()` helper assertion.
  If the implementation of the "No Matches Message" display becomes volatile
  for the near future (hopefully it won't), it'll be helpful to consolidate
  the logic for locating this message to a single place for our tests.