feat: Enable Combobox Typeahead Searching to find `option`s with spaces
The native `<select>` element has this behavior, and this behavior provides
a convenient UX in cases where some `option`s are more lengthy. Thus, we've
decided to add support for this feature.

SIDENOTE: We also increased Playwright's Test Retries in CI for the Typeahead
features. On VERY rare occasions, there's flakiness in CI for the time-based
tests in this Describe/Feature Block. The hope with this change is that having
5 retries will reduce that "resultant flakiness" down to virtually 0 in CI.
Locally, the tests work fine. (I only experienced a flaky failure for a
time-based test once, when my computer was exceptionally busy.)