feat: Enable `SelectEnhancer` to be setup directly with component parts
In the past, `SelectEnhancer` expected to have a `<select>` as a child.
Now, we allow rendering the `<select-enhancer>` with a `<combobox-field>`
and a `<combobox-listbox>` as a child. However, the right number (and
proper arrangement) of elements must be present.

Due to how this setup process works, we're back to requiring that
`ComboboxListbox` is the `nextElementSibling` of `ComboboxField`.
There doesn't seem to be a simple way around this.

There are more notes that we need to make regarding the order in which
the `connectedCallback()`s of the various Web Components are called.
We also need to add support for using `<select>` with child instances
of the `ComboboxField`/`ComboboxListbox` components. (This is already
supported when the `combobox` + `listbox` combination is supplied
directly.) However, we thought it best to push a commit now due to
some complexities that we encountered while trying to support this
feature.