fix: Remove text selection from `combobox` when it loses focus
I suspect that the component was keeping text selection when `blur`red with
a Keyboard in the past because the browser considers the element distinct
from regular text-editable elements. Since it's a "regular" element that's
`[contenteditable]`, it's kinda like highlighting the text of a `<div>`,
except it's focusable. So the browser didn't see a need to change the text
selection on `blur` on its own. Thus, we have to do this manually.