chore: Add example of integrating Combobox component with Form Observer
Our other tests should already be enough. But just for the sake of
sanity, we wanted to add some tests showing that the component can
easily integrate with other reliable Form Libraries.

Note that at this point, we have so many tests that it seems like
some of them are becoming flaky (particularly the ones that do
really wild things like dynamically creating new Custom Element
Constructors). We aren't sure where the flakiness comes from,
but all tests ARE indeed passing. Maybe we can find a way to
remove the flakiness in the future.