feat: Make `ComboboxOption` `selected` state changes safe pre-mounting
This turned out to be necessary for certain kinds of JS Frameworks like
React, which may set the properties/attributes of a Custom Element BEFORE
connecting it to the DOM. We now take an approach similar to the native
`<select>` element: Developers can update the selected state of `option`s
however they want, but once the `option`s are connected to the DOM,
we take over and make sure that the selected state of all `option`s is
corrected. If the `<select>` element does this, it seems reasonable for
us to do this as well.

The native `<select>` element technically performs this correction once
the `option`s are appended to it. By contrast, our component performs
this correction logic once it's connected to the DOM. Practically
speaking, this shouldn't really matter since the data/state only matters
once the component is connected to the DOM. But if it became of interest,
we could explore the idea of running the correction logic pre-DOM-connection
(probably unnecessary/useless and not worthwhile).

As a side note, we updated some of our internal development notes.
There are thoughts that we didn't document a while ago that we thought
should be documented now, along with some thoughts that ARE relevant
with this commit.

Also committed some `index.html` updates that we forgot to supply earlier.