chore: Add notes about some project decisions and future ideas
These really should've been added earlier, but we were "in the zone"
and really wanted to make sure we were cranking out good features for
the `ComboboxField` component before we forgot said ideas.

Now that we have a clearer picture of what we want to do, we decided
to document the thoughts that weren't previously documented.