refactor: Import example styles with `<link>`s instead of with JS
This makes our example code usable in non-Vite contexts for
rapid prototyping or quick debugging. Motivated by the debugging
and experimenting that we did in order to understand the order
in which Custom Elements are constructed and have their
`connectedCallback()`s called. (Same for the conversion from
`main.ts` to `main.js` too.)