feat: Synchronize `listbox` and `option` IDs when `combobox` ID changes
In JS frameworks, you never know if/when people intend to update the IDs
of their DOM elements, so supporting this feature is important. Altering
the `listbox` and/or `option` IDs is still illegal and will break the
component. However, with this update, users can at least change all of
the IDs simply be changing the `combobox` ID. And the IDs of the other
elements will be clear and predictable based on the `combobox`'s ID.