chore: Document Pros/Cons of using attributeChangedCallback vs. Observers
While attempting to refactor our `MutationObserver`s for `aria-expanded`
and `aria-activedescendant` to an `attributeChangedCallback()` implementation,
we discovered that it's probably more effort than it's worth at the moment.
Since we're hoping to release soon, the refactor isn't a priority. Maybe
we could try it in the future? But there are clear pros/cons to both
approaches, even if `attributeChangedCallback()` is much to be desired for
[memory] performance purposes.