refactor: Place all `MutationObserver` callbacks in the same location
This just helps for grouping our logic better, and it removes a lot of
clutter from the beginning of the component. Note that we had to use
`bind`ing because the actual `MutationObserver`s are declared at the
beginning of the class definition. Thus, we had to make methods that
exist _before_ the observers were created, and which could be bound
to the callbacks passed to those observers. This is slightly more
memory-efficient than arrow functions anyway, so this is fine.